workflow: delete unwired ML subsystem (Issue #1081)
Remove five unreachable ML files that compiled into every binary
importing features/workflow without being called by any wired path:

  ml_integration.go    — MLEnhancedEngine wrapper (unwired)
  ml_logging.go        — MLLogger type (unwired)
  ml_export.go         — MLDataExporter type (unwired)
  ml_export_test.go    — tests for deleted exporter
  ml_logging_test.go   — tests for deleted logger (used testify/mock,
                         violating CFGMS no-mocks standard)

Pre-deletion grep confirmed zero callers outside these five files:
  grep -r "MLEnhancedEngine|MLLogger|MLDataExporter" . → no matches

PerformanceMetrics struct was defined in ml_logging.go but consumed by
performance_collector.go (an in-scope survivor). Moved the struct there
so the collector compiles. Added performance_collector_test.go to
restore test coverage for PerformanceCollector and PerformanceMetrics;
tests use real runtime package calls, no mocks.

Also fixes a pre-existing gofmt alignment violation in
features/siem/stream_processor_test.go caught by golangci-lint.

Fixes #1081
